// src/api/generated/models/paginatedAttendanceResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * ECD Backend API
 * Early Childhood Development management system API.
 * OpenAPI spec version: 1.0
 */
import type { AttendanceResponseDto } from './attendanceResponseDto';

export interface PaginatedAttendanceResponseDto {
  items: AttendanceResponseDto[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}