// src/api/generated/models/absentReason.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * ECD Backend API
 * Early Childhood Development management system API.
 * OpenAPI spec version: 1.0
 */

export type AbsentReason = typeof AbsentReason[keyof typeof AbsentReason];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AbsentReason = {
  sick: 'sick',
  family: 'family',
  transport: 'transport',
  weather: 'weather',
  unknown: 'unknown',
  other: 'other',
} as const;